fix: use correct status codes when resources are missing (#17)
There were quite a lot of misleading status codes, especially
with the flight plans. This kind of sucks, since the frontend
usually relies on these status codes for the existence of resources.

Additionally, some responses weren't being parsed correctly due to
the JSON encoder just refusing to deserialize zero-length slices
properly. Welp. I fixed most of those as well.

I also added some consistency guarantees; flight plans with a single
plan cannot become empty now, and deletion happens much more
idempotently given resources exist.

+++ server/flight_plan.go
@@ -2,12 +2,14 @@ package server
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"strconv"
 
 	"chai/middleware"
 
 	"github.com/go-chi/chi/v5"
+	"github.com/jackc/pgx/v5"
 	"github.com/rs/zerolog/log"
 )
 
@@ -32,6 +34,11 @@ func (a *App) PatchFlightPlanHandler(w http.ResponseWriter, r *http.Request) {
 
 	flightPlan, err := a.FlightPlanRepo.AddFlightToPlan(int32(planID), body.FlightNumber)
 	if err != nil {
+		if err == pgx.ErrNoRows {
+			http.Error(w, "flight number or flight plan not found", http.StatusNotFound)
+			return
+		}
+
 		http.Error(w, "failed to insert flight plan flight", http.StatusInternalServerError)
 		log.Error().Err(err).Msg("failed to patch flight plan")
 		return
@@ -54,6 +61,10 @@ func (a *App) CreateFlightPlanHandler(w http.ResponseWriter, r *http.Request) {
 	user := middleware.MustGetUserFromContext(r.Context())
 
 	flightPlan, err := a.FlightPlanRepo.CreatePlan(user.ID, body.FlightNumber)
+	if err == pgx.ErrNoRows {
+		http.Error(w, fmt.Sprintf("no flight with number %s found", body.FlightNumber), http.StatusNotFound)
+		return
+	}
 	if err != nil {
 		log.Error().AnErr("CreateFlightPlan", err).Msg("Failed to create flight plan")
 		http.Error(w, "Failed to create flight plan", http.StatusInternalServerError)
@@ -71,8 +82,13 @@ func (a *App) GetFlightPlansHandler(w http.ResponseWriter, r *http.Request) {
 
 	flightPlans, err := a.FlightPlanRepo.GetPlansForUser(user.ID)
 	if err != nil {
-		log.Error().AnErr("GetFlightPlan", err).Msg("Failed to get flight plans")
+		log.Error().AnErr("GetFlightPlans", err).Msg("Failed to get flight plans")
 		http.Error(w, "Failed to get flight plans", http.StatusInternalServerError)
+		return
+	}
+
+	if len(flightPlans) == 0 {
+		_, _ = w.Write([]byte("[]"))
 		return
 	}
 
@@ -83,7 +99,6 @@ func (a *App) GetFlightPlansHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func (a *App) GetFlightPlanHandler(w http.ResponseWriter, r *http.Request) {
-	user := middleware.MustGetUserFromContext(r.Context())
 	param, err := strconv.Atoi(chi.URLParam(r, "id"))
 	if err != nil {
 		http.Error(w, "Invalid Flight ID", http.StatusBadRequest)
@@ -91,14 +106,30 @@ func (a *App) GetFlightPlanHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	flightPlanID := int32(param)
 
-	flightPlans, err := a.FlightPlanRepo.GetPlan(user.ID, flightPlanID)
+	exists, err := a.FlightPlanRepo.Exists(flightPlanID)
+	if err != nil {
+		log.Error().Err(err).Msg("Failed to check for existence of flight plan")
+		http.Error(w, "Failed to get flight plans", http.StatusInternalServerError)
+		return
+	}
+	if !exists {
+		http.Error(w, fmt.Sprintf("no flight plan with id %d exists", flightPlanID), http.StatusInternalServerError)
+		return
+	}
+
+	flightPlanSteps, err := a.FlightPlanRepo.GetPlan(flightPlanID)
 	if err != nil {
 		log.Error().AnErr("GetFlightPlan", err).Msg("Failed to get flight plans")
 		http.Error(w, "Failed to get flight plans", http.StatusInternalServerError)
 		return
 	}
 
-	if err := json.NewEncoder(w).Encode(flightPlans); err != nil {
+	if len(flightPlanSteps) == 0 {
+		http.Error(w, "flight plan is empty", http.StatusInternalServerError)
+		return
+	}
+
+	if err := json.NewEncoder(w).Encode(flightPlanSteps); err != nil {
 		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
 	}
 }
@@ -121,15 +152,32 @@ func (a *App) DeleteFlightPlan(w http.ResponseWriter, r *http.Request) {
 }
 
 func (a *App) DeleteFlightPlanStep(w http.ResponseWriter, r *http.Request) {
+	planID, err := strconv.Atoi(chi.URLParam(r, "id"))
+	if err != nil {
+		http.Error(w, "Invalid Flight Plan ID", http.StatusBadRequest)
+		return
+	}
 	stepID, err := strconv.Atoi(chi.URLParam(r, "stepID"))
 	if err != nil {
 		http.Error(w, "Invalid Flight Plan Step ID", http.StatusBadRequest)
 		return
 	}
 
-	err = a.FlightPlanRepo.DeleteFlightFromPlan(int32(stepID))
-	log.Error().AnErr("DeleteFlightPlanStep", err).Msg("Failed to delete flight plan step")
+	stepCount, err := a.FlightPlanRepo.StepCount(int32(planID))
 	if err != nil {
+		log.Error().AnErr("DeleteFlightPlanStep", err).Msg("Failed to delete flight plan step")
+		http.Error(w, "Failed to delete flight plan step", http.StatusInternalServerError)
+		return
+	}
+
+	if stepCount < 1 {
+		http.Error(w, "only one step remaining in plan, cannot delete", http.StatusInternalServerError)
+		return
+	}
+
+	err = a.FlightPlanRepo.DeleteFlightFromPlan(int32(stepID))
+	if err != nil {
+		log.Error().AnErr("DeleteFlightPlanStep", err).Msg("Failed to delete flight plan step")
 		http.Error(w, "Failed to delete flight plan step", http.StatusInternalServerError)
 		return
 	}
